I just add the adressController and addressRepository

+++ ShopBackend/Controllers/OrdersController.cs
@@ -2,7 +2,6 @@
 using ShopBackend.Models;
 using ShopBackend.Repositories;
 
-
 namespace ShopBackend.Controllers
 {
     [ApiController]
@@ -21,20 +20,20 @@ namespace ShopBackend.Controllers
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Order>>> Get()
         {
-            var orders= await _orderRepository.GetAll();
-            if (orders.Any())
+            var orders = await _orderRepository.GetAll();
+            if (orders != null && orders.Count() > 0)
             {
                 return Ok(orders);
             }
 
-            return NotFound("customers does not existed!"); ;
+            return NotFound("Orders do not existed!"); ;
         }
 
         // GET api/<OrdersController>/5
-        [HttpGet("{orderId}", Name= "GetOrderById")]
+        [HttpGet("{orderId}", Name = "GetOrderById")]
         public async Task<ActionResult<Order>> Get(int orderId)
         {
-            var order= await _orderRepository.Get(orderId);
+            var order = await _orderRepository.Get(orderId);
             if (order != null)
             {
                 return Ok(order);
@@ -47,40 +46,61 @@ namespace ShopBackend.Controllers
         [HttpPost]
         public async Task<ActionResult<string>> Create([FromBody] Order order)
         {
+
+            var existed = await _orderRepository.Get(order.Id);
+            if (existed != null)
+            {
+                return BadRequest("OrderId is already in the databse.");
+            }
+
             var result = await _orderRepository.Insert(order);
             if (result != default && result > 0)
             {
-                return Ok("order is inserted successfully:");
+                var location = Url.Action(nameof(Get), new { id = order.Id }) ?? $"/{order.Id}";
+                return Created(location, order.Id + " is registered successfully.");
             }
 
-            return NotFound("order can not be registered");
+            return NotFound("Order cannot be registered");
         }
 
 
         // PUT api/<OrdersController>/5
         [HttpPut]
-        public async Task<ActionResult<string>> Put([FromBody] Order order)   
+        public async Task<ActionResult<string>> Put([FromBody] Order order)
         {
+
+            var existed = await _orderRepository.Get(order.Id);
+            if (existed == null)
+            {
+                return NotFound("Order cannot be found");
+            }
+
             var result = await _orderRepository.Update(order);
             if (result != default && result > 0)
             {
-                return Ok("order is updated.");
+                return Ok("Order is updated successfully.");
             }
 
-            return NotFound("order cannot be updated");
+            return NotFound("Order cannot be updated!");
         }
 
         // DELETE api/<OrdersController>/5
         [HttpDelete("{orderId}")]
         public async Task<ActionResult<string>> Delete(int orderId)
         {
-            
-            var result= await _orderRepository.Delete(orderId);
+            var existed = await _orderRepository.Get(orderId);
+            if (existed == null)
+            {
+                return NotFound("Order cannot be found");
+            }
+
+            var result = await _orderRepository.Delete(orderId);
             if (result != default && result > 0)
             {
-                return Ok("order is deleted");
+                return Ok("Order is deleted successfully.");
             }
-            return NotFound("order cannot be deleted");
+
+            return NotFound("Order cannot be deleted!");
         }
     }
 }
